requirements.txt
flask
mysql-connector-python
azure-storage-blob
pandas